# helpers/swagger.yaml
openapi: 3.0.0
info:
  title: The System MySQL Database
  description: The System MySQL Database
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
...
  /accounts:
    get:
      summary: Get a list of all accounts
      description: Retrieve list of all accounts.
      operationId: getAll
      security:
        - bearerAuth: []
      responses:
        '200':
          description: An array of all accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    accountNumber:
                      type: string
                      example: 'AAA001'
                    accountName:
                      type: string
                      example: 'Disney'
                    openDate:
                      type: string
                      minLength: 8
                      maxLength: 8
                      format: date
                      pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                      example: 20190816
                      description: 'Date the account was opened.'
                    debtorAge:
                      type: integer
                      example: 30
                    paymentTermDays:
                      type: integer
                      example: 60
                    creditLimit:
                      type: number
                      example: 1500.00
                    totalBalance:
                      type: number
                      example: 1000.00
                    amountDue:
                      type: number
                      example: 1500.00
                    currentBalance:
                      type: number
                      example: 1000.00
                    days30:
                      type: number
                      example: 30.00
                    days60:
                      type: number
                      example: 60.00
                    days90:
                      type: number
                      example: 90.00
                    days120:
                      type: number
                      example: 120.00
                    days150:
                      type: number
                      example: 150.00
                    days180:
                      type: number
                      example: 180.00
                    days180Over:
                      type: number
                      example: 200.00
                    paymentMethod:
                      type: string
                      example: 'EFT'
                    paymentDueDate:
                      type: integer
                      example: 2
                    debitOrderDate:
                      type: integer
                      example: 2
                    lastPaymentDate:
                      type: string
                      minLength: 8
                      maxLength: 8
                      format: date
                      pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                      example: 20190816
                    lastPaymentAmount:
                      type: number
                      example: 2016.00
                    lastPTPDate:
                      type: string
                      minLength: 8
                      maxLength: 8
                      format: date
                      pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                      example: 20190816
                    lastPTPAmount:
                      type: number
                      example: 1016.00
                    accountNotes:
                      type: string
                      example: 'Customer can be shouty'
                    accountStatus:
                      type: string
                      example: 'Open'
                    arg:
                      type: string
                      example: 'A1'
                    createdAt:
                      type: string
                      example: '2020-05-05T09:12:57.848Z'
                    createdBy:
                      type: string
                      example: 'mickey@example.com'
                    updatedAt:
                      type: string
                      example: '2020-05-08T03:11:21.553Z'
                    updatedBy:
                      type: string
                      example: 'mickey@example.com'
                    f_customerRefNo:
                      type: string
                      example: 'AAA001'
                  required:
                    - accountNumber
        '401':
          $ref: '#/components/responses/UnauthorisedError'
    post:
      summary: Create a new account
      description: Restricted to admin users.
      operationId: create
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  accountNumber:
                    type: string
                    example: 'AAA001'
                  accountName:
                    type: string
                    example: 'Disney'
                  openDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                    description: 'Date the account was opened.'
                  debtorAge:
                    type: integer
                    example: 30
                  paymentTermDays:
                    type: integer
                    example: 60
                  creditLimit:
                    type: number
                    example: 1500.00
                  totalBalance:
                    type: number
                    example: 1000.00
                  amountDue:
                    type: number
                    example: 1500.00
                  currentBalance:
                    type: number
                    example: 1000.00
                  days30:
                    type: number
                    example: 30.00
                  days60:
                    type: number
                    example: 60.00
                  days90:
                    type: number
                    example: 90.00
                  days120:
                    type: number
                    example: 120.00
                  days150:
                    type: number
                    example: 150.00
                  days180:
                    type: number
                    example: 180.00
                  days180Over:
                    type: number
                    example: 200.00
                  paymentMethod:
                    type: string
                    example: 'EFT'
                  paymentDueDate:
                    type: integer
                    example: 2
                  debitOrderDate:
                    type: integer
                    example: 2
                  lastPaymentDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                  lastPaymentAmount:
                    type: number
                    example: 2016.00
                  lastPTPDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                  lastPTPAmount:
                    type: number
                    example: 1016.00
                  accountNotes:
                    type: string
                    example: 'Customer can be shouty'
                  accountStatus:
                    type: string
                    example: 'Open'
                  arg:
                    type: string
                    example: 'A1'
                  createdAt:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  createdBy:
                    type: string
                    example: 'mickey@example.com'
                  updatedAt:
                    type: string
                    example: '2020-05-08T03:11:21.553Z'
                  updatedBy:
                    type: string
                    example: 'mickey@example.com'
                  f_customerId:
                    type: string
                    example: 'AAA101'
              required:
                - accountNumber
                - accountName
                - createdBy
      responses:
        '200':
          description: Account created successfully. The account number of the new account is returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountNumber:
                    type: string
                    example: 'AAA001'
        '400':
          description: Account name is already on the books
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Account name 'Disney' is already on the books"
        '401':
          $ref: '#/components/responses/UnauthorisedError'
  /accounts/{accountNumber}:
    parameters:
      - in: path
        name: accountNumber
        description: Account number
        required: true
        example: 'AAA001'
        schema:
          type: string
    get:
      summary: Get a single account by account number
      description: Retrieve a single account by account number.
      operationId: getById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Details of the specified account
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountNumber:
                    type: string
                    example: 'AAA001'
                  accountName:
                    type: string
                    example: 'Disney'
                  openDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                    description: 'Date the account was opened.'
                  debtorAge:
                    type: integer
                    example: 30
                  paymentTermDays:
                    type: integer
                    example: 60
                  creditLimit:
                    type: number
                    example: 1500.00
                  totalBalance:
                    type: number
                    example: 1000.00
                  amountDue:
                    type: number
                    example: 1500.00
                  currentBalance:
                    type: number
                    example: 1000.00
                  days30:
                    type: number
                    example: 30.00
                  days60:
                    type: number
                    example: 60.00
                  days90:
                    type: number
                    example: 90.00
                  days120:
                    type: number
                    example: 120.00
                  days150:
                    type: number
                    example: 150.00
                  days180:
                    type: number
                    example: 180.00
                  days180Over:
                    type: number
                    example: 200.00
                  paymentMethod:
                    type: string
                    example: 'EFT'
                  paymentDueDate:
                    type: integer
                    example: 2
                  debitOrderDate:
                    type: integer
                    example: 2
                  lastPaymentDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                  lastPaymentAmount:
                    type: number
                    example: 2016.00
                  lastPTPDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                  lastPTPAmount:
                    type: number
                    example: 1016.00
                  accountNotes:
                    type: string
                    example: 'Customer can be shouty'
                  accountStatus:
                    type: string
                    example: 'Open'
                  arg:
                    type: string
                    example: 'A1'
                  createdAt:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  createdBy:
                    type: string
                    example: 'mickey@example.com'
                  updatedAt:
                    type: string
                    example: '2020-05-08T03:11:21.553Z'
                  updatedBy:
                    type: string
                    example: 'mickey@example.com'
                  f_customerRefNo:
                    type: string
                    example: 'AAA101'
              required:
                - accountNumber
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorisedError'
    put:
      summary: Update an account
      description: Update details of an account.
      operationId: update
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accountNumber:
                  type: string
                  example: 'AAA001'
                accountName:
                  type: string
                  example: 'Disney'
                openDate:
                  type: string
                  minLength: 8
                  maxLength: 8
                  format: date
                  pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                  example: 20190816
                  description: 'Date the account was opened.'
                debtorAge:
                  type: integer
                  example: 30
                paymentTermDays:
                  type: integer
                  example: 60
                creditLimit:
                  type: number
                  example: 1500.00
                totalBalance:
                  type: number
                  example: 1000.00
                amountDue:
                  type: number
                  example: 1500.00
                currentBalance:
                  type: number
                  example: 1000.00
                days30:
                  type: number
                  example: 30.00
                days60:
                  type: number
                  example: 60.00
                days90:
                  type: number
                  example: 90.00
                days120:
                  type: number
                  example: 120.00
                days150:
                  type: number
                  example: 150.00
                days180:
                  type: number
                  example: 180.00
                days180Over:
                  type: number
                  example: 200.00
                paymentMethod:
                  type: string
                  example: 'EFT'
                paymentDueDate:
                  type: integer
                  example: 2
                debitOrderDate:
                  type: integer
                  example: 2
                lastPaymentDate:
                  type: string
                  minLength: 8
                  maxLength: 8
                  format: date
                  pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                  example: 20190816
                lastPaymentAmount:
                  type: number
                  example: 2016.00
                lastPTPDate:
                  type: string
                  minLength: 8
                  maxLength: 8
                  format: date
                  pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                  example: 20190816
                lastPTPAmount:
                  type: number
                  example: 1016.00
                accountNotes:
                  type: string
                  example: 'Customer can be shouty'
                accountStatus:
                  type: string
                  example: 'Open'
                arg:
                  type: string
                  example: 'A1'
                createdAt:
                  type: string
                  example: '2020-05-05T09:12:57.848Z'
                createdBy:
                  type: string
                  example: 'mickey@example.com'
                updatedAt:
                  type: string
                  example: '2020-05-08T03:11:21.553Z'
                updatedBy:
                  type: string
                  example: 'mickey@example.com'
                f_customerRefNo:
                  type: string
                  example: 'AAA101'
            required:
              - accountNumber
              - updatedBy
      responses:
        '200':
          description: Account updated successfully. The details of the updated account are returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accountNumber:
                    type: string
                    example: 'AAA001'
                  accountName:
                    type: string
                    example: 'Disney'
                  openDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                    description: 'Date the account was opened.'
                  debtorAge:
                    type: integer
                    example: 30
                  paymentTermDays:
                    type: integer
                    example: 60
                  creditLimit:
                    type: number
                    example: 1500.00
                  totalBalance:
                    type: number
                    example: 1000.00
                  amountDue:
                    type: number
                    example: 1500.00
                  currentBalance:
                    type: number
                    example: 1000.00
                  days30:
                    type: number
                    example: 30.00
                  days60:
                    type: number
                    example: 60.00
                  days90:
                    type: number
                    example: 90.00
                  days120:
                    type: number
                    example: 120.00
                  days150:
                    type: number
                    example: 150.00
                  days180:
                    type: number
                    example: 180.00
                  days180Over:
                    type: number
                    example: 200.00
                  paymentMethod:
                    type: string
                    example: 'EFT'
                  paymentDueDate:
                    type: integer
                    example: 2
                  debitOrderDate:
                    type: integer
                    example: 2
                  lastPaymentDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                  lastPaymentAmount:
                    type: number
                    example: 2016.00
                  lastPTPDate:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                  lastPTPAmount:
                    type: number
                    example: 1016.00
                  accountNotes:
                    type: string
                    example: 'Customer can be shouty'
                  accountStatus:
                    type: string
                    example: 'Open'
                  arg:
                    type: string
                    example: 'A1'
                  createdAt:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  createdBy:
                    type: string
                    example: 'mickey@example.com'
                  updatedAt:
                    type: string
                    example: '2020-05-08T03:11:21.553Z'
                  updatedBy:
                    type: string
                    example: 'mickey@example.com'
                  f_customerRefNo:
                    type: string
                    example: 'AAA101'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorisedError'
    delete:
      summary: Delete an account
      description: Delete an account.
      operationId: delete
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Account deleted successfully'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorisedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorisedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Unauthorised'
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Not Found'